SELC-5122 refactor: moved check organization function in dedicated class
